test(piper): move global patching test to dedicated file
